refactor: centralize protected artwork type contract
